+++ WolvenKit.App/ViewModels/Editor/Tools/ImportExportViewModel.cs
@@ -6,25 +6,24 @@ using System.Threading.Tasks;
 using System.Windows.Input;
 using System.Linq;
 using Catel;
+using Catel.MVVM;
 using Catel.Services;
-using CP77.CR2W;
 using DynamicData;
 using ReactiveUI;
-using WolvenKit.Common;
 using WolvenKit.Common.Model.Arguments;
 using WolvenKit.Common.Services;
 using WolvenKit.Functionality.Commands;
 using WolvenKit.Functionality.Controllers;
 using WolvenKit.Functionality.Services;
-using WolvenKit.Models;
 using WolvenKit.RED4.CR2W.Archive;
 using ModTools = WolvenKit.Modkit.RED4.ModTools;
-using ObservableObject = Catel.Data.ObservableObject;
 
 namespace WolvenKit.ViewModels.Editor
 {
     public class ImportExportViewModel : ToolViewModel
     {
+        #region fields
+
         private const string s_selectedInGrid = "Selected in Grid";
 
         /// <summary>
@@ -36,90 +35,6 @@ namespace WolvenKit.ViewModels.Editor
         /// Identifies the caption string used for this tool window.
         /// </summary>
         public const string ToolTitle = "Import Export Tool";
-
-        /// <summary>
-        /// Private Importable Items
-        /// </summary>
-        private readonly ReadOnlyObservableCollection<ImportableItemViewModel> _importableItems;
-
-        /// <summary>
-        /// Public Importable Items
-        /// </summary>
-        public ReadOnlyObservableCollection<ImportableItemViewModel> ImportableItems => _importableItems;
-
-        /// <summary>
-        /// Private Exportable Items
-        /// </summary>
-        private readonly ReadOnlyObservableCollection<ExportableItemViewModel> _exportableItems;
-
-        /// <summary>
-        /// Public Exportable items.
-        /// </summary>
-        public ReadOnlyObservableCollection<ExportableItemViewModel> ExportableItems => _exportableItems;
-
-        /// <summary>
-        /// Selected Export Item
-        /// </summary>
-        public ExportableItemViewModel SelectedExport { get; set; }
-
-        /// <summary>
-        /// Selected Import Item
-        /// </summary>
-        public ImportableItemViewModel SelectedImport { get; set; }
-
-        public ObservableCollection<ImportableItemViewModel> SelectedImports { get; set; }
-
-        /// <summary>
-        /// Selected object , returns a Importable/Exportable ItemVM based on "IsImportsSelected"
-        /// </summary>
-        public ImportExportItemViewModel SelectedObject => IsImportsSelected ? SelectedImport : SelectedExport;
-
-        public bool? IsHeaderChecked { get; set; }
-
-        /// <summary>
-        /// Private NameOf Selected Item in Grid.
-        /// </summary>
-        private string _CurrentSelectionInGridName;
-
-        /// <summary>
-        /// Private Last Selected Item, Used for Selection Lock.
-        /// </summary>
-        private ImportExportItemViewModel lastselected;
-
-        /// <summary>
-        /// Lock Selection of items in grid.
-        /// </summary>
-        public bool SelectionLocked { get; set; } = false;
-
-        /// <summary>
-        /// Returns the name of current selected object in Import/Export Grid.
-        /// </summary>
-        public string CurrentSelectedInGridName
-        {
-            get
-            {
-                if (SelectedObject != null)
-                {
-                    if (!SelectionLocked)
-                    {
-                        lastselected = SelectedObject;
-                        return SelectedObject.Name;
-                    }
-                    else
-                    {
-                        if (lastselected == null)
-                        { return ""; }
-                        else
-                        {
-                            return lastselected.Name;
-                        }
-                    }
-                }
-                else
-                { return ""; }
-            }
-            set { _CurrentSelectionInGridName = value; }
-        }
 
         /// <summary>
         /// Private Readonly ModTools
@@ -151,7 +66,27 @@ namespace WolvenKit.ViewModels.Editor
         /// </summary>
         private readonly IGameControllerFactory _gameController;
 
-        
+        /// <summary>
+        /// Private NameOf Selected Item in Grid.
+        /// </summary>
+        private string _CurrentSelectionInGridName;
+
+        /// <summary>
+        /// Private Last Selected Item, Used for Selection Lock.
+        /// </summary>
+        private ImportExportItemViewModel lastselected;
+
+        /// <summary>
+        /// Private Importable Items
+        /// </summary>
+        private readonly ReadOnlyObservableCollection<ImportableItemViewModel> _importableItems;
+
+        /// <summary>
+        /// Private Exportable Items
+        /// </summary>
+        private readonly ReadOnlyObservableCollection<ExportableItemViewModel> _exportableItems;
+
+        #endregion
 
         /// <summary>
         /// Import Export ViewModel Constructor
@@ -187,10 +122,8 @@ namespace WolvenKit.ViewModels.Editor
 
             SetupToolDefaults();
 
-            SelectedImports = new ObservableCollection<ImportableItemViewModel>();
-
-            ProcessAllCommand = new RelayCommand(ExecuteProcessAll, CanProcessAll);
-            ProcessSelectedCommand = new RelayCommand(ExecuteProcessSelected, CanProcessSelected);
+            ProcessAllCommand = new TaskCommand(ExecuteProcessAll, CanProcessAll);
+            ProcessSelectedCommand = new TaskCommand(ExecuteProcessSelected, CanProcessSelected);
 
             CopyArgumentsTemplateToCommand = new DelegateCommand<string>(ExecuteCopyArgumentsTemplateTo, CanCopyArgumentsTemplateTo);
 
@@ -213,15 +146,79 @@ namespace WolvenKit.ViewModels.Editor
                 .Subscribe();
         }
 
+        #region properties
+
+       
+
+        /// <summary>
+        /// Public Importable Items
+        /// </summary>
+        public ReadOnlyObservableCollection<ImportableItemViewModel> ImportableItems => _importableItems;
+
+        /// <summary>
+        /// Public Exportable items.
+        /// </summary>
+        public ReadOnlyObservableCollection<ExportableItemViewModel> ExportableItems => _exportableItems;
+
+        /// <summary>
+        /// Selected Export Item
+        /// </summary>
+        public ExportableItemViewModel SelectedExport { get; set; }
+
+        /// <summary>
+        /// Selected Import Item
+        /// </summary>
+        public ImportableItemViewModel SelectedImport { get; set; }
+
+        /// <summary>
+        /// Selected object , returns a Importable/Exportable ItemVM based on "IsImportsSelected"
+        /// </summary>
+        public ImportExportItemViewModel SelectedObject => IsImportsSelected ? SelectedImport : SelectedExport;
+
+        public bool? IsHeaderChecked { get; set; }
+
+        /// <summary>
+        /// Lock Selection of items in grid.
+        /// </summary>
+        public bool SelectionLocked { get; set; } = false;
+
+        /// <summary>
+        /// Returns the name of current selected object in Import/Export Grid.
+        /// </summary>
+        public string CurrentSelectedInGridName
+        {
+            get
+            {
+                if (SelectedObject != null)
+                {
+                    if (!SelectionLocked)
+                    {
+                        lastselected = SelectedObject;
+                        return SelectedObject.Name;
+                    }
+                    else
+                    {
+                        if (lastselected == null)
+                        { return ""; }
+                        else
+                        {
+                            return lastselected.Name;
+                        }
+                    }
+                }
+                else
+                { return ""; }
+            }
+            set => _CurrentSelectionInGridName = value;
+        }
+
+
         /// <summary>
         /// Is Import Selected, if false Export is default.
         /// </summary>
         public bool IsImportsSelected { get; set; }
 
-        /// <summary>
-        /// Process all in Import / Export Grid Command.
-        /// </summary>
-        public ICommand ProcessAllCommand { get; private set; }
+        #endregion
 
         public ICommand CopyArgumentsTemplateToCommand { get; private set; }
 
@@ -272,6 +269,11 @@ namespace WolvenKit.ViewModels.Editor
         public bool IsProcessing { get; set; } = false;
 
         /// <summary>
+        /// Process all in Import / Export Grid Command.
+        /// </summary>
+        public ICommand ProcessAllCommand { get; private set; }
+
+        /// <summary>
         /// Can Process all Bool
         /// </summary>
         /// <returns>bool</returns>
@@ -280,7 +282,7 @@ namespace WolvenKit.ViewModels.Editor
         /// <summary>
         /// Execute Process all in Import / Export Grid Command.
         /// </summary>
-        private void ExecuteProcessAll()
+        private async Task ExecuteProcessAll()
         {
             IsProcessing = true;
 
@@ -288,14 +290,14 @@ namespace WolvenKit.ViewModels.Editor
             {
                 foreach (var item in ImportableItems)
                 {
-                    ImportSingle(item);
+                    await ImportSingle(item);
                 }
             }
             else
             {
                 foreach (var item in ExportableItems)
                 {
-                    ExportSingle(item);
+                    await ExportSingle(item);
                 }
             }
             IsProcessing = false;
@@ -305,14 +307,14 @@ namespace WolvenKit.ViewModels.Editor
         /// Import Single item
         /// </summary>
         /// <param name="item"></param>
-        private void ImportSingle(ImportableItemViewModel item)
+        private async Task ImportSingle(ImportableItemViewModel item)
         {
             var proj = _projectManager.ActiveProject;
             var fi = new FileInfo(item.FullName);
             if (fi.Exists)
             {
                 var settings = new GlobalImportArgs().Register(item.Properties as ImportArgs);
-                _modTools.Import(fi, settings, new DirectoryInfo(proj.ModDirectory));
+                await Task.Run(() =>  _modTools.Import(fi, settings, new DirectoryInfo(proj.ModDirectory)));
             }
         }
 
@@ -320,7 +322,7 @@ namespace WolvenKit.ViewModels.Editor
         /// Export Single Item
         /// </summary>
         /// <param name="item"></param>
-        private void ExportSingle(ExportableItemViewModel item)
+        private async Task ExportSingle(ExportableItemViewModel item)
         {
             var proj = _projectManager.ActiveProject;
             var fi = new FileInfo(item.FullName);
@@ -334,7 +336,9 @@ namespace WolvenKit.ViewModels.Editor
                 }
 
                 var settings = new GlobalExportArgs().Register(item.Properties as ExportArgs);
-                _modTools.Export(fi, settings, new DirectoryInfo(proj.ModDirectory), new DirectoryInfo(proj.RawDirectory));
+                await Task.Run(() => _modTools.Export(fi, settings,
+                    new DirectoryInfo(proj.ModDirectory),
+                    new DirectoryInfo(proj.RawDirectory)));
             }
         }
 
@@ -342,8 +346,6 @@ namespace WolvenKit.ViewModels.Editor
         /// Process selected in Import / Export Grid Command
         /// </summary>
         public ICommand ProcessSelectedCommand { get; private set; }
-
-        
 
         /// <summary>
         /// Can Process Selected Bool.
@@ -354,21 +356,21 @@ namespace WolvenKit.ViewModels.Editor
         /// <summary>
         /// Execute Process selected in Import / Export Grid Command
         /// </summary>
-        private void ExecuteProcessSelected()
+        private async Task ExecuteProcessSelected()
         {
             IsProcessing = true;
             if (IsImportsSelected)
             {
                 foreach (var item in ImportableItems.Where(_ => _.IsChecked))
                 {
-                    ImportSingle(item);
+                    await ImportSingle(item);
                 }
             }
             else
             {
                 foreach (var item in ExportableItems.Where(_ => _.IsChecked))
                 {
-                    ExportSingle(item);
+                    await ExportSingle(item);
                 }
             }
             IsProcessing = false;
@@ -394,99 +396,5 @@ namespace WolvenKit.ViewModels.Editor
         /// Setup Tool defaults for tool window.
         /// </summary>
         private void SetupToolDefaults() => ContentId = ToolContentId;
-    }
-
-    /// <summary>
-    /// ImportExportItem ViewModel
-    /// </summary>
-    public abstract class ImportExportItemViewModel : ObservableObject
-    {
-        /// <summary>
-        /// BaseFile "FileModel"
-        /// </summary>
-        protected FileModel BaseFile { get; set; }
-
-        /// <summary>
-        /// Properties
-        /// </summary>
-        public ImportExportArgs Properties
-        {
-            get => _properties;
-            set
-            {
-                if (_properties != value)
-                {
-                    var oldValue = _properties;
-                    _properties = value;
-                    RaisePropertyChanged(() => Properties, oldValue, value);
-                    RaisePropertyChanged(() => ExportTaskIdentifier);
-                }
-            }
-        }
-
-        private ImportExportArgs _properties;
-
-        public string ExportTaskIdentifier => Properties.ToString();
-
-        public string Extension => BaseFile.GetExtension();
-        public string FullName => BaseFile.FullName;
-        public string Name => BaseFile.Name;
-
-        public bool IsChecked { get; set; }
-
-        public EExportState ExportState => BaseFile.IsImportable ? EExportState.Importable : EExportState.Exportable;
-    }
-
-    public class ImportableItemViewModel : ImportExportItemViewModel
-    {
-        public ImportableItemViewModel(FileModel model)
-        {
-            BaseFile = model;
-            Properties = DecideImportOptions(model);
-        }
-
-        private ImportArgs DecideImportOptions(FileModel model)
-        {
-            _ = Enum.TryParse(model.GetExtension(), out ERawFileFormat rawFileFormat);
-
-            return rawFileFormat switch
-            {
-                ERawFileFormat.tga => new XbmImportArgs(),
-                ERawFileFormat.dds => new XbmImportArgs(),
-                ERawFileFormat.fbx => new CommonImportArgs(),
-                ERawFileFormat.glb => new MeshImportArgs(),
-                ERawFileFormat.gltf => new MeshImportArgs(),
-                _ => new CommonImportArgs()
-            };
-        }
-    }
-
-    public class ExportableItemViewModel : ImportExportItemViewModel
-    {
-        public ExportableItemViewModel(FileModel model)
-        {
-            BaseFile = model;
-            Properties = DecideExportOptions(model);
-        }
-
-        private ExportArgs DecideExportOptions(FileModel model)
-        {
-            _ = Enum.TryParse(model.GetExtension(), out ECookedFileFormat fileFormat);
-
-            return fileFormat switch
-            {
-                ECookedFileFormat.mesh => new MeshExportArgs(),
-                ECookedFileFormat.xbm => new XbmExportArgs(),
-                ECookedFileFormat.wem => new WemExportArgs(),
-                ECookedFileFormat.csv => new CommonExportArgs(),
-                //ECookedFileFormat.json => new CommonExportArgs(),
-                ECookedFileFormat.mlmask => new CommonExportArgs(),
-                ECookedFileFormat.cubemap => new CommonExportArgs(),
-                ECookedFileFormat.envprobe => new CommonExportArgs(),
-                ECookedFileFormat.texarray => new CommonExportArgs(),
-                ECookedFileFormat.morphtarget => new MorphTargetExportArgs(),
-                _ => throw new ArgumentOutOfRangeException()
-            };
-        }
     }
 }
